Handle Firebase swapping empty arrays with `null`

+++ firebase-web/src/main/java/io/spine/web/firebase/FirebaseRecord.java
@@ -83,7 +83,7 @@ final class FirebaseRecord {
     /**
      * Writes this record to the given {@link FirebaseDatabase} in a single transaction
      * (i.e. in a single batch).
-     * 
+     *
      * <p>Receiving data from Spine and writing it to database are both performed asynchronously.
      */
     void storeTransactionallyTo(FirebaseDatabase database) {
@@ -96,7 +96,7 @@ final class FirebaseRecord {
      *
      * @return an integer number of records
      */
-    int getCount() {
+    long getCount() {
         CountConsumer countConsumer = new CountConsumer();
         queryResponse.thenAccept(countConsumer);
         return countConsumer.getValue();
@@ -107,7 +107,7 @@ final class FirebaseRecord {
      */
     private static class CountConsumer implements Consumer<QueryResponse> {
 
-        private int value;
+        private long value;
 
         @Override
         public void accept(QueryResponse response) {
@@ -117,7 +117,7 @@ final class FirebaseRecord {
         /**
          * @return the count of messages in the consumed response
          */
-        public int getValue() {
+        public long getValue() {
             return value;
         }
     }
